fix(DropdownMenu): rename demo
