+++ src/dropdown-menu/demos/disabled.vue
@@ -1,0 +1,6 @@
+<template>
+  <t-dropdown-menu>
+    <t-dropdown-item disabled label="禁用菜单" />
+    <t-dropdown-item disabled label="禁用菜单" />
+  </t-dropdown-menu>
+</template>
